pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocsv>=1.3.2",
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.15",
    "asyncpg>=0.30.0",
    "chromadb>=1.0.20",
    "langchain>=0.3.27",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "sentence-transformers>=5.1.0",
    "sqlalchemy>=2.0.43",
]
